requirements.txt.txt
requests 
beautifulsoup4
python-whois 
dnspython